Factor out escape_uri_attribute One fewer "magic" type system lie! Explicitly use the `->forceAttribute()` api. This is the proper way to tell lies and will be caught when xhp attributes become sound.
